make sure the total for BPH becomes an int

+++ DrinkUp/BalancedHelper.py
@@ -130,7 +130,7 @@ class BalancedPaymentsHelper:
 		
 		try:
 			hold = buyer.hold(
-				amount=round(float(order.grand_total), 2)*100, #this needs to be in pennies
+				amount=int(round(float(order.grand_total), 2)*100), #this needs to be in pennies
 				description= buyer.name + ' at ' + order.bar.venue.name,
 				source_uri=source_uri
 			)
